userspace: split process.go by concern (Refs #4421)
pkg/dataplane/userspace/process.go had grown to 1208 lines holding
~32 unrelated top-level declarations spanning process lifecycle,
control-socket plumbing, the status loop, NAPI/neighbor prewarm, and
the link-cycle rebind machinery. This is a pure code-motion split from
the refactor backlog (#4421, codex-173 #3): no declaration is added,
removed, renamed, or edited — each is moved verbatim (doc comment and
body byte-for-byte) into a sibling file grouped by concern.

New files (all package userspace, same directory):

  process.go            process lifecycle: ensureProcessLocked,
                        tuneSocketBuffers, findBinary, stopLocked,
                        configEqual, StartFIBSync
  process_control.go    control-socket request plumbing:
                        MaxControlRequestBytes, deadline consts,
                        controlRoundtripDeadline, requestDetailedLocked,
                        sessionSocketPath, requestSessionSync,
                        requestLocked
  process_status.go     status/health loop + snapshot publish:
                        syncSnapshotLocked, ensureStatusLoopLocked,
                        statusLoop, shouldStandbyNeighborPrewarmLocked
  process_napi.go       NAPI bootstrap + neighbor prewarm probes:
                        bootstrapNAPIQueues{,Async}Locked,
                        proactiveNeighborResolve{,Async}{,Locked},
                        neighborProbeTarget, sendICMPProbe*,
                        sendUDPProbeForNAPI
  process_linkcycle.go  ctrl enable/disable + link-cycle rebind:
                        {disable,reEnable}UserspaceCtrlLocked,
                        DisableAndStopHelper, PrepareLinkCycle,
                        linkCycleRebindSleep, NotifyLinkCycle

Per-file imports were pruned to only what each concern uses; all five
files are gofmt-clean.

Validation: the split was proven byte-identical with a go/ast decl
extractor — the sorted top-level production decl-NAME set has an empty
diff both ways (32 == 32) and every decl body is byte-for-byte identical
between origin/master's single file and the union of the five new files
(body-multiset identical). go build ./..., go vet
./pkg/dataplane/userspace/..., and go test ./pkg/dataplane/userspace/...
are all green.
